@@ -1,93 +1,125 @@
 {
-  "name": "bitconch-web3j",
-  "version": "0.0.0-development",
-  "description": "Javascript API for bitconch chain",
+  "name": "@bitconch/bitconch-web3j",
+  "version": "1.1.4",
+  "description": "Bitconch Javascript API",
   "keywords": [
     "api",
     "blockchain",
     "javascript",
-    "web3",
     "bitconch"
   ],
+  "testnetDefaultChannel": "edge",
   "license": "MIT",
-  "author": "Bitconch dev",
-  "homepage": "https://github.com/caesarchad/bitconch-web3j#readme",
+  "author": "Bitconch Dev",
+  "homepage": "https://bitconch.io/",
   "repository": {
     "type": "git",
-    "url": "https://github.com/caesarchad/bitconch-web3j"
+    "url": "https://github.com/bitconch/bitconch-web3j.git"
   },
   "bugs": {
-    "url": "https://github.com/caesarchad/bitconch-web3j/issues"
+    "url": "http://github.com/bitconch/bitconch-web3j.git/issues"
   },
   "publishConfig": {
     "access": "public"
   },
   "main": "lib/index.cjs.js",
   "module": "lib/index.esm.js",
+  "bin": {
+    "mini_server": "bin/mini_server.sh"
+  },
+  "files": [
+    "/bin",
+    "/doc",
+    "/lib",
+    "/module.flow.js",
+    "/src"
+  ],
   "scripts": {
-    "clean": "rimraf ./coverage ./lib",
-    "dev": "cross-env NODE_ENV=development rollup -c",
+    "bpf-sdk:install": "set -ex; npm run bpf-sdk:fetch; npm run bpf-sdk:clean",
+    "bpf-sdk:fetch": "curl http://solana-sdk.s3.amazonaws.com/$npm_package_testnetDefaultChannel/bpf-sdk.tar.bz2 -o bpf-sdk.tar.bz2",
+    "bpf-sdk:remove-symlinks": "find bpf-sdk -type l -print -exec cp {} {}.tmp \\; -exec mv {}.tmp {} \\;",
+    "bpf-sdk:clean": "rm bpf-sdk.tar.bz2",
     "build": "cross-env NODE_ENV=production rollup -c",
-    "doc": "esdoc; node -p '\"\\nDocumentation coverage: \" + require(\"./doc/coverage.json\").coverage'",
+    "clean": "rimraf ./coverage ./lib",
+    "codecov": "set -ex; npm run test:cover; cat ./coverage/lcov.info | codecov",
+    "dev": "cross-env NODE_ENV=development rollup -c",
+    "doc": "set -ex; esdoc; node -p '\"\\nDocumentation coverage: \" + require(\"./doc/coverage.json\").coverage'",
     "doc:watch": "watch 'npm run doc' . --wait=1 --ignoreDirectoryPattern=/doc/",
-    "test": "cross-env NODE_ENV=test jest --useStderr",
-    "test:live": "cross-env NODE_ENV=test DOITLIVE=1 jest --useStderr",
-    "test:watch": "cross-env NODE_ENV=test jest --watch --useStderr",
-    "test:cover": "cross-env NODE_ENV=test jest --coverage --useStderr",
-    "codecov": "npm run test:cover && cat ./coverage/lcov.info | codecov",
-    "flow": "flow-typed install jest@22 && flow",
-    "flow:watch": "watch 'flow' . --wait=1 --ignoreDirectoryPattern=/doc/",
-    "lint": "eslint .",
+    "examples": "set -ex; for example in examples/*.js; do node $example; done",
+    "flow": "set -ex; flow stop; flow-typed install jest@24; flow",
+    "flow:stop": "flow stop",
+    "flow:watch": "flow stop; watch 'flow' . --wait=1 --ignoreDirectoryPattern=/doc/",
+    "install": "npm run bpf-sdk:install",
+    "lint": "set -ex; npm run pretty; eslint .",
     "lint:fix": "npm run lint -- --fix",
     "lint:watch": "watch 'npm run lint:fix' . --wait=1 --ignoreDirectoryPattern=/doc/",
-    "ok": "npm run lint && npm run flow && npm run test && npm run doc",
-    "prepare": "npm run clean && npm run ok && npm run build",
-    "examples": "set -ex; for example in examples/*.js; do node $example; done",
-    "re": "semantic-release --repository-url git@github.com:bitconch/bitconch-web3j.git"
+    "localnet:down": "bin/localnet.sh down",
+    "localnet:logs": "bin/localnet.sh logs -f",
+    "localnet:up": "bin/localnet.sh up $npm_package_testnetDefaultChannel",
+    "localnet:update": "bin/localnet.sh update $npm_package_testnetDefaultChannel",
+    "ok": "set -ex; npm run lint; npm run flow; npm run test; npm run doc",
+    "prepare": "set -ex; npm run clean; npm run bpf-sdk:install; npm run build",
+    "pretty": "prettier --write '{,{examples,src,test}/**/}*.js'",
+    "re": "semantic-release --repository-url git@github.com:bitconch/bitconch-web3j.git",
+    "test": "cross-env NODE_ENV=test jest --useStderr",
+    "test:cover": "cross-env NODE_ENV=test jest --coverage --useStderr",
+    "test:live": "cross-env NODE_ENV=test TEST_LIVE=1 jest --useStderr",
+    "test:watch": "cross-env NODE_ENV=test jest --watch --useStderr"
   },
   "dependencies": {
-    "babel-runtime": "^6.26.0",
+    "@babel/runtime": "^7.3.1",
     "bn.js": "^4.11.8",
     "bs58": "^4.0.1",
-    "jayson": "^2.0.6",
+    "buffer-layout": "^1.2.0",
+    "esdoc-inject-style-plugin": "^1.0.0",
+    "jayson": "^3.0.1",
+    "mz": "^2.7.0",
     "node-fetch": "^2.2.0",
+    "rpc-websockets": "^4.3.3",
     "superstruct": "^0.6.0",
-    "tweetnacl": "^1.0.0"
+    "tweetnacl": "^1.0.0",
+    "ws": "^6.1.0"
   },
   "devDependencies": {
-    "babel-core": "6.26.3",
-    "babel-eslint": "10.0.0",
-    "babel-plugin-module-resolver": "3.1.1",
-    "babel-plugin-transform-runtime": "^6.23.0",
-    "babel-preset-env": "1.7.0",
-    "babel-preset-flow": "6.23.0",
-    "babel-preset-stage-0": "6.24.1",
+    "@babel/core": "^7.3.3",
+    "@babel/plugin-proposal-class-properties": "^7.3.3",
+    "@babel/plugin-proposal-function-bind": "^7.2.0",
+    "@babel/plugin-transform-runtime": "^7.2.0",
+    "@babel/preset-env": "^7.3.1",
+    "@babel/preset-flow": "^7.0.0",
+    "babel-eslint": "10.0.1",
+    "babel-plugin-module-resolver": "3.2.0",
     "codecov": "^3.0.4",
     "cross-env": "5.2.0",
-    "enzyme": "3.6.0",
+    "elfy": "^0.1.0",
+    "enzyme": "3.9.0",
     "esdoc": "^1.1.0",
     "esdoc-ecmascript-proposal-plugin": "^1.0.0",
     "esdoc-flow-type-plugin": "^1.1.0",
     "esdoc-importpath-plugin": "^1.0.2",
     "esdoc-standard-plugin": "^1.0.0",
-    "eslint": "5.6.0",
-    "eslint-plugin-import": "2.14.0",
-    "eslint-plugin-jest": "21.22.1",
-    "flow-bin": "0.81.0",
+    "eslint": "5.16.0",
+    "eslint-config-prettier": "^4.0.0",
+    "eslint-plugin-import": "2.17.2",
+    "eslint-plugin-jest": "22.5.1",
+    "eslint-plugin-prettier": "^3.0.0",
+    "flow-bin": "0.97.0",
     "flow-typed": "2.5.1",
-    "fs-file-tree": "1.0.6",
-    "jest": "23.6.0",
-    "rimraf": "2.6.2",
-    "rollup": "0.66.2",
-    "rollup-plugin-babel": "3.0.3",
-    "rollup-plugin-commonjs": "9.1.8",
-    "rollup-plugin-json": "^3.0.0",
+    "fs-file-tree": "1.0.7",
+    "jest": "24.7.1",
+    "marked": "^0.6.0",
+    "prettier": "^1.14.3",
+    "rimraf": "2.6.3",
+    "rollup": "1.11.3",
+    "rollup-plugin-babel": "^4.3.2",
+    "rollup-plugin-commonjs": "9.3.4",
+    "rollup-plugin-json": "^4.0.0",
     "rollup-plugin-node-builtins": "^2.1.2",
     "rollup-plugin-node-globals": "^1.2.1",
-    "rollup-plugin-node-resolve": "3.4.0",
-    "rollup-plugin-replace": "2.0.0",
-    "rollup-plugin-uglify": "3.0.0",
-    "semantic-release": "^15.9.16",
+    "rollup-plugin-node-resolve": "4.2.3",
+    "rollup-plugin-replace": "2.2.0",
+    "rollup-plugin-terser": "^4.0.4",
+    "semantic-release": "^15.13.17",
     "watch": "^1.0.2"
   }
 }
