@@ -1,37 +1,56 @@
 // @flow
 
 import assert from 'assert';
+import {parse as urlParse, format as urlFormat} from 'url';
 import fetch from 'node-fetch';
 import jayson from 'jayson/lib/client/browser';
 import {struct} from 'superstruct';
+import {Client as RpcWebSocketClient} from 'rpc-websockets';
 
-import {Transaction} from './transaction';
+import {DEFAULT_TICKS_PER_SLOT, NUM_TICKS_PER_SECOND} from './timing';
 import {PublicKey} from './publickey';
+import {Transaction} from './transaction';
+import {sleep} from './util/sleep';
+import type {Blockhash} from './blockhash';
 import type {Account} from './account';
-import type {TransactionSignature, TransactionId} from './transaction';
+import type {TransactionSignature} from './transaction';
 
 type RpcRequest = (methodName: string, args: Array<any>) => any;
 
-function createRpcRequest(url): RpcRequest {
-  const server = jayson(
-    async (request, callback) => {
-      const options = {
-        method: 'POST',
-        body: request,
-        headers: {
-          'Content-Type': 'application/json',
-        }
-      };
+/**
+ * Information describing a cluster node
+ *
+ * @typedef {Object} ContactInfo
+ * @property {string} id Unique identifier of the node
+ * @property {string} gossip Gossip network address for the node
+ * @property {string} tpu TPU network address for the node (null if not available)
+ * @property {string|null} rpc JSON RPC network address for the node (null if not available)
+ */
+type ContactInfo = {
+  id: string,
+  gossip: string,
+  tpu: string | null,
+  rpc: string | null,
+};
 
-      try {
-        const res = await fetch(url, options);
-        const text = await res.text();
-        callback(null, text);
-      } catch (err) {
-        callback(err);
-      }
+function createRpcRequest(url): RpcRequest {
+  const server = jayson(async (request, callback) => {
+    const options = {
+      method: 'POST',
+      body: request,
+      headers: {
+        'Content-Type': 'application/json',
+      },
+    };
+
+    try {
+      const res = await fetch(url, options);
+      const text = await res.text();
+      callback(null, text);
+    } catch (err) {
+      callback(err);
     }
-  );
+  });
 
   return (method, args) => {
     return new Promise((resolve, reject) => {
@@ -46,7 +65,6 @@ function createRpcRequest(url): RpcRequest {
   };
 }
 
-
 /**
  * Expected JSON RPC response for the "getBalance" message
  */
@@ -57,7 +75,6 @@ const GetBalanceRpcResult = struct({
   result: 'number?',
 });
 
-
 /**
  * @private
  */
@@ -67,7 +84,7 @@ function jsonRpcResult(resultDescription: any) {
     struct({
       jsonrpc: jsonRpcVersion,
       id: 'string',
-      error: 'any'
+      error: 'any',
     }),
     struct({
       jsonrpc: jsonRpcVersion,
@@ -78,16 +95,42 @@ function jsonRpcResult(resultDescription: any) {
   ]);
 }
 
+/**
+ * @private
+ */
+const AccountInfoResult = struct({
+  executable: 'boolean',
+  owner: 'array',
+  // lamports: 'number',
+  dif: 'number',
+  data: 'array',
+});
 
 /**
  * Expected JSON RPC response for the "getAccountInfo" message
  */
-const GetAccountInfoRpcResult = jsonRpcResult({
-  program_id: 'array',
-  tokens: 'number',
-  userdata: 'array',
+const GetAccountInfoRpcResult = jsonRpcResult(AccountInfoResult);
+
+/***
+ * Expected JSON RPC response for the "accountNotification" message
+ */
+const AccountNotificationResult = struct({
+  subscription: 'number',
+  result: AccountInfoResult,
 });
 
+/**
+ * @private
+ */
+const ProgramAccountInfoResult = struct(['string', AccountInfoResult]);
+
+/***
+ * Expected JSON RPC response for the "programNotification" message
+ */
+const ProgramAccountNotificationResult = struct({
+  subscription: 'number',
+  result: ProgramAccountInfoResult,
+});
 
 /**
  * Expected JSON RPC response for the "confirmTransaction" message
@@ -95,14 +138,33 @@ const GetAccountInfoRpcResult = jsonRpcResult({
 const ConfirmTransactionRpcResult = jsonRpcResult('boolean');
 
 /**
+ * Expected JSON RPC response for the "getSlotLeader" message
+ */
+const GetSlotLeader = jsonRpcResult('string');
+
+/**
+ * Expected JSON RPC response for the "getClusterNodes" message
+ */
+const GetClusterNodes = jsonRpcResult(
+  struct.list([
+    struct({
+      id: 'string',
+      gossip: 'string',
+      tpu: struct.union(['null', 'string']),
+      rpc: struct.union(['null', 'string']),
+    }),
+  ]),
+);
+
+/**
  * Expected JSON RPC response for the "getSignatureStatus" message
  */
-const GetSignatureStatusRpcResult = jsonRpcResult(struct.enum([
-  'Confirmed',
-  'SignatureNotFound',
-  'ProgramRuntimeError',
-  'GenericFailure',
-]));
+const GetSignatureStatusRpcResult = jsonRpcResult(
+  struct.union([
+    'null',
+    struct.union([struct({Ok: 'null'}), struct({Err: 'object'})]),
+  ]),
+);
 
 /**
  * Expected JSON RPC response for the "getTransactionCount" message
@@ -110,14 +172,9 @@ const GetSignatureStatusRpcResult = jsonRpcResult(struct.enum([
 const GetTransactionCountRpcResult = jsonRpcResult('number');
 
 /**
- * Expected JSON RPC response for the "getLastId" message
+ * Expected JSON RPC response for the "getRecentBlockhash" message
  */
-const GetLastId = jsonRpcResult('string');
-
-/**
- * Expected JSON RPC response for the "getFinality" message
- */
-const GetFinalityRpcResult = jsonRpcResult('number');
+const GetRecentBlockhash = jsonRpcResult('string');
 
 /**
  * Expected JSON RPC response for the "requestAirdrop" message
@@ -127,34 +184,106 @@ const RequestAirdropRpcResult = jsonRpcResult('string');
 /**
  * Expected JSON RPC response for the "sendTransaction" message
  */
-const SendTokensRpcResult = jsonRpcResult('string');
+const SendTransactionRpcResult = jsonRpcResult('string');
 
 /**
  * Information describing an account
  *
  * @typedef {Object} AccountInfo
- * @property {number} tokens Number of tokens assigned to the account
- * @property {PublicKey} programId Identifier of the program assigned to the account
- * @property {?Buffer} userdata Optional userdata assigned to the account
+//  * @property {number} lamports Number of lamports assigned to the account
+ * @property {number} dif Number of dif assigned to the account
+ * @property {PublicKey} owner Identifier of the program that owns the account
+ * @property {?Buffer} data Optional data assigned to the account
+ * @property {boolean} executable `true` if this account's data contains a loaded program
  */
 type AccountInfo = {
-  tokens: number,
-  programId: PublicKey,
-  userdata: Buffer,
-}
+  executable: boolean,
+  owner: PublicKey,
+  // lamports: number,
+  dif: number,
+  data: Buffer,
+};
 
 /**
- * Possible signature status values
+ * Account information identified by pubkey
  *
- * @typedef {string} SignatureStatus
+ * @typedef {Object} KeyedAccountInfo
+ * @property {PublicKey} accountId
+ * @property {AccountInfo} accountInfo
  */
-type SignatureStatus = 'Confirmed' | 'SignatureNotFound' | 'ProgramRuntimeError' | 'GenericFailure';
+type KeyedAccountInfo = {
+  accountId: PublicKey,
+  accountInfo: AccountInfo,
+};
+
+/**
+ * Callback function for account change notifications
+ */
+export type AccountChangeCallback = (accountInfo: AccountInfo) => void;
+
+/**
+ * @private
+ */
+type AccountSubscriptionInfo = {
+  publicKey: string, // PublicKey of the account as a base 58 string
+  callback: AccountChangeCallback,
+  subscriptionId: null | number, // null when there's no current server subscription id
+};
+
+/**
+ * Callback function for program account change notifications
+ */
+export type ProgramAccountChangeCallback = (
+  keyedAccountInfo: KeyedAccountInfo,
+) => void;
+
+/**
+ * @private
+ */
+type ProgramAccountSubscriptionInfo = {
+  programId: string, // PublicKey of the program as a base 58 string
+  callback: ProgramAccountChangeCallback,
+  subscriptionId: null | number, // null when there's no current server subscription id
+};
+
+/**
+ * Signature status: Success
+ *
+ * @typedef {Object} SignatureSuccess
+ */
+export type SignatureSuccess = {|
+  Ok: null,
+|};
+
+/**
+ * Signature status: TransactionError
+ *
+ * @typedef {Object} TransactionError
+ */
+export type TransactionError = {|
+  Err: Object,
+|};
 
 /**
  * A connection to a fullnode JSON RPC endpoint
  */
 export class Connection {
   _rpcRequest: RpcRequest;
+  _rpcWebSocket: RpcWebSocketClient;
+  _rpcWebSocketConnected: boolean = false;
+
+  _blockhashInfo: {
+    recentBlockhash: Blockhash | null,
+    seconds: number,
+    transactionSignatures: Array<string>,
+  };
+  _disableBlockhashCaching: boolean = false;
+  _accountChangeSubscriptions: {[number]: AccountSubscriptionInfo} = {};
+  _accountChangeSubscriptionCounter: number = 0;
+  _programAccountChangeSubscriptions: {
+    [number]: ProgramAccountSubscriptionInfo,
+  } = {};
+  _programAccountChangeSubscriptionCounter: number = 0;
 
   /**
    * Establish a JSON RPC connection
@@ -162,20 +291,45 @@ export class Connection {
    * @param endpoint URL to the fullnode JSON RPC endpoint
    */
   constructor(endpoint: string) {
-    if (typeof endpoint !== 'string') {
-      throw new Error('Connection endpoint not specified');
+    let url = urlParse(endpoint);
+
+    this._rpcRequest = createRpcRequest(url.href);
+    this._blockhashInfo = {
+      recentBlockhash: null,
+      seconds: -1,
+      transactionSignatures: [],
+    };
+
+    url.protocol = url.protocol === 'https:' ? 'wss:' : 'ws:';
+    url.host = '';
+    url.port = String(Number(url.port) + 1);
+    if (url.port === '1') {
+      url.port = url.protocol === 'wss:' ? '8901' : '8900';
     }
-    this._rpcRequest = createRpcRequest(endpoint);
+    this._rpcWebSocket = new RpcWebSocketClient(urlFormat(url), {
+      autoconnect: false,
+      max_reconnects: Infinity,
+    });
+    this._rpcWebSocket.on('open', this._wsOnOpen.bind(this));
+    this._rpcWebSocket.on('error', this._wsOnError.bind(this));
+    this._rpcWebSocket.on('close', this._wsOnClose.bind(this));
+    this._rpcWebSocket.on(
+      'accountNotification',
+      this._wsOnAccountNotification.bind(this),
+    );
+    this._rpcWebSocket.on(
+      'programNotification',
+      this._wsOnProgramAccountNotification.bind(this),
+    );
   }
 
   /**
    * Fetch the balance for the specified public key
    */
   async getBalance(publicKey: PublicKey): Promise<number> {
-    const unsafeRes = await this._rpcRequest(
-      'getBalance',
-      [publicKey.toBase58()]
-    );
+    const unsafeRes = await this._rpcRequest('getBalance', [
+      publicKey.toBase58(),
+    ]);
     const res = GetBalanceRpcResult(unsafeRes);
     if (res.error) {
       throw new Error(res.error.message);
@@ -188,10 +342,9 @@ export class Connection {
    * Fetch all the account info for the specified public key
    */
   async getAccountInfo(publicKey: PublicKey): Promise<AccountInfo> {
-    const unsafeRes = await this._rpcRequest(
-      'getAccountInfo',
-      [publicKey.toBase58()]
-    );
+    const unsafeRes = await this._rpcRequest('getAccountInfo', [
+      publicKey.toBase58(),
+    ]);
     const res = GetAccountInfoRpcResult(unsafeRes);
     if (res.error) {
       throw new Error(res.error.message);
@@ -201,9 +354,11 @@ export class Connection {
     assert(typeof result !== 'undefined');
 
     return {
-      tokens: result.tokens,
-      programId: new PublicKey(result.program_id),
-      userdata: Buffer.from(result.userdata),
+      executable: result.executable,
+      owner: new PublicKey(result.owner),
+      // lamports: result.lamports,
+      dif: result.dif,
+      data: Buffer.from(result.data),
     };
   }
 
@@ -211,10 +366,7 @@ export class Connection {
    * Confirm the transaction identified by the specified signature
    */
   async confirmTransaction(signature: TransactionSignature): Promise<boolean> {
-    const unsafeRes = await this._rpcRequest(
-      'confirmTransaction',
-      [signature]
-    );
+    const unsafeRes = await this._rpcRequest('confirmTransaction', [signature]);
     const res = ConfirmTransactionRpcResult(unsafeRes);
     if (res.error) {
       throw new Error(res.error.message);
@@ -224,9 +376,37 @@ export class Connection {
   }
 
   /**
-   * Fetch the current transaction count of the network
+   * Fetch the current slot leader of the cluster
    */
-  async getSignatureStatus(signature: TransactionSignature): Promise<SignatureStatus> {
+  async getSlotLeader(): Promise<string> {
+    const unsafeRes = await this._rpcRequest('getSlotLeader', []);
+    const res = GetSlotLeader(unsafeRes);
+    if (res.error) {
+      throw new Error(res.error.message);
+    }
+    assert(typeof res.result !== 'undefined');
+    return res.result;
+  }
+
+  /**
+   * Return the list of nodes that are currently participating in the cluster
+   */
+  async getClusterNodes(): Promise<Array<ContactInfo>> {
+    const unsafeRes = await this._rpcRequest('getClusterNodes', []);
+    const res = GetClusterNodes(unsafeRes);
+    if (res.error) {
+      throw new Error(res.error.message);
+    }
+    assert(typeof res.result !== 'undefined');
+    return res.result;
+  }
+
+  /**
+   * Fetch the current transaction count of the cluster
+   */
+  async getSignatureStatus(
+    signature: TransactionSignature,
+  ): Promise<SignatureSuccess | TransactionError | null> {
     const unsafeRes = await this._rpcRequest('getSignatureStatus', [signature]);
     const res = GetSignatureStatusRpcResult(unsafeRes);
     if (res.error) {
@@ -236,9 +416,8 @@ export class Connection {
     return res.result;
   }
 
-
   /**
-   * Fetch the current transaction count of the network
+   * Fetch the current transaction count of the cluster
    */
   async getTransactionCount(): Promise<number> {
     const unsafeRes = await this._rpcRequest('getTransactionCount', []);
@@ -251,11 +430,11 @@ export class Connection {
   }
 
   /**
-   * Fetch the identifier to the latest transaction on the network
+   * Fetch a recent blockhash from the cluster
    */
-  async getLastId(): Promise<TransactionId> {
-    const unsafeRes = await this._rpcRequest('getLastId', []);
-    const res = GetLastId(unsafeRes);
+  async getRecentBlockhash(): Promise<Blockhash> {
+    const unsafeRes = await this._rpcRequest('getRecentBlockhash', []);
+    const res = GetRecentBlockhash(unsafeRes);
     if (res.error) {
       throw new Error(res.error.message);
     }
@@ -264,23 +443,16 @@ export class Connection {
   }
 
   /**
-   * Return the current network finality time in millliseconds
+   * Request an allocation of lamports to the specified account
    */
-  async getFinality(): Promise<number> {
-    const unsafeRes = await this._rpcRequest('getFinality', []);
-    const res = GetFinalityRpcResult(unsafeRes);
-    if (res.error) {
-      throw new Error(res.error.message);
-    }
-    assert(typeof res.result !== 'undefined');
-    return Number(res.result);
-  }
-
-  /**
-   * Request an allocation of tokens to the specified account
-   */
-  async requestAirdrop(to: PublicKey, amount: number): Promise<TransactionSignature> {
-    const unsafeRes = await this._rpcRequest('requestAirdrop', [to.toBase58(), amount]);
+  async requestAirdrop(
+    to: PublicKey,
+    amount: number,
+  ): Promise<TransactionSignature> {
+    const unsafeRes = await this._rpcRequest('requestAirdrop', [
+      to.toBase58(),
+      amount,
+    ]);
     const res = RequestAirdropRpcResult(unsafeRes);
     if (res.error) {
       throw new Error(res.error.message);
@@ -292,18 +464,333 @@ export class Connection {
   /**
    * Sign and send a transaction
    */
-  async sendTransaction(from: Account, transaction: Transaction): Promise<TransactionSignature> {
-    transaction.lastId = await this.getLastId();
-    transaction.sign(from);
+  async sendTransaction(
+    transaction: Transaction,
+    ...signers: Array<Account>
+  ): Promise<TransactionSignature> {
+    for (;;) {
+      // Attempt to use a recent blockhash for up to 30 seconds
+      const seconds = new Date().getSeconds();
+      if (
+        this._blockhashInfo.recentBlockhash != null &&
+        this._blockhashInfo.seconds < seconds + 30
+      ) {
+        transaction.recentBlockhash = this._blockhashInfo.recentBlockhash;
+        transaction.sign(...signers);
+        if (!transaction.signature) {
+          throw new Error('!signature'); // should never happen
+        }
+
+        // If the signature of this transaction has not been seen before with the
+        // current recentBlockhash, all done.
+        const signature = transaction.signature.toString();
+        if (!this._blockhashInfo.transactionSignatures.includes(signature)) {
+          this._blockhashInfo.transactionSignatures.push(signature);
+          if (this._disableBlockhashCaching) {
+            this._blockhashInfo.seconds = -1;
+          }
+          break;
+        }
+      }
+
+      // Fetch a new blockhash
+      let attempts = 0;
+      const startTime = Date.now();
+      for (;;) {
+        const recentBlockhash = await this.getRecentBlockhash();
+
+        if (this._blockhashInfo.recentBlockhash != recentBlockhash) {
+          this._blockhashInfo = {
+            recentBlockhash,
+            seconds: new Date().getSeconds(),
+            transactionSignatures: [],
+          };
+          break;
+        }
+        if (attempts === 50) {
+          throw new Error(
+            `Unable to obtain a new blockhash after ${Date.now() -
+              startTime}ms`,
+          );
+        }
+
+        // Sleep for approximately half a slot
+        await sleep((500 * DEFAULT_TICKS_PER_SLOT) / NUM_TICKS_PER_SECOND);
+
+        ++attempts;
+      }
+    }
 
     const wireTransaction = transaction.serialize();
-    const unsafeRes = await this._rpcRequest('sendTransaction', [[...wireTransaction]]);
-    const res = SendTokensRpcResult(unsafeRes);
+    return await this.sendRawTransaction(wireTransaction);
+  }
+
+  /**
+   * @private
+   */
+  async fullnodeExit(): Promise<boolean> {
+    const unsafeRes = await this._rpcRequest('fullnodeExit', []);
+    const res = jsonRpcResult('boolean')(unsafeRes);
+    if (res.error) {
+      throw new Error(res.error.message);
+    }
+    assert(typeof res.result !== 'undefined');
+    return res.result;
+  }
+
+  /**
+   * Send a transaction that has already been signed and serialized into the
+   * wire format
+   */
+  async sendRawTransaction(
+    rawTransaction: Buffer,
+  ): Promise<TransactionSignature> {
+    const unsafeRes = await this._rpcRequest('sendTransaction', [
+      [...rawTransaction],
+    ]);
+    const res = SendTransactionRpcResult(unsafeRes);
     if (res.error) {
       throw new Error(res.error.message);
     }
     assert(typeof res.result !== 'undefined');
     assert(res.result);
     return res.result;
+  }
+
+  /**
+   * @private
+   */
+  _wsOnOpen() {
+    this._rpcWebSocketConnected = true;
+    this._updateSubscriptions();
+  }
+
+  /**
+   * @private
+   */
+  _wsOnError(err: Error) {
+    console.log('ws error:', err.message);
+  }
+
+  /**
+   * @private
+   */
+  _wsOnClose(code: number, message: string) {
+    // 1000 means _rpcWebSocket.close() was called explicitly
+    if (code !== 1000) {
+      console.log('ws close:', code, message);
+    }
+    this._rpcWebSocketConnected = false;
+  }
+
+  /**
+   * @private
+   */
+  async _updateSubscriptions() {
+    const accountKeys = Object.keys(this._accountChangeSubscriptions).map(
+      Number,
+    );
+    const programKeys = Object.keys(
+      this._programAccountChangeSubscriptions,
+    ).map(Number);
+    if (accountKeys.length === 0 && programKeys.length === 0) {
+      this._rpcWebSocket.close();
+      return;
+    }
+
+    if (!this._rpcWebSocketConnected) {
+      for (let id of accountKeys) {
+        this._accountChangeSubscriptions[id].subscriptionId = null;
+      }
+      for (let id of programKeys) {
+        this._programAccountChangeSubscriptions[id].subscriptionId = null;
+      }
+      this._rpcWebSocket.connect();
+      return;
+    }
+
+    for (let id of accountKeys) {
+      const {subscriptionId, publicKey} = this._accountChangeSubscriptions[id];
+      if (subscriptionId === null) {
+        try {
+          this._accountChangeSubscriptions[
+            id
+          ].subscriptionId = await this._rpcWebSocket.call('accountSubscribe', [
+            publicKey,
+          ]);
+        } catch (err) {
+          console.log(
+            `accountSubscribe error for ${publicKey}: ${err.message}`,
+          );
+        }
+      }
+    }
+    for (let id of programKeys) {
+      const {
+        subscriptionId,
+        programId,
+      } = this._programAccountChangeSubscriptions[id];
+      if (subscriptionId === null) {
+        try {
+          this._programAccountChangeSubscriptions[
+            id
+          ].subscriptionId = await this._rpcWebSocket.call('programSubscribe', [
+            programId,
+          ]);
+        } catch (err) {
+          console.log(
+            `programSubscribe error for ${programId}: ${err.message}`,
+          );
+        }
+      }
+    }
+  }
+
+  /**
+   * @private
+   */
+  _wsOnAccountNotification(notification: Object) {
+    const res = AccountNotificationResult(notification);
+    if (res.error) {
+      throw new Error(res.error.message);
+    }
+
+    const keys = Object.keys(this._accountChangeSubscriptions).map(Number);
+    for (let id of keys) {
+      const sub = this._accountChangeSubscriptions[id];
+      if (sub.subscriptionId === res.subscription) {
+        const {result} = res;
+        assert(typeof result !== 'undefined');
+
+        sub.callback({
+          executable: result.executable,
+          owner: new PublicKey(result.owner),
+          // lamports: result.lamports,
+          dif: result.dif,
+          data: Buffer.from(result.data),
+        });
+        return true;
+      }
+    }
+  }
+
+  /**
+   * Register a callback to be invoked whenever the specified account changes
+   *
+   * @param publickey Public key of the account to monitor
+   * @param callback Function to invoke whenever the account is changed
+   * @return subscription id
+   */
+  onAccountChange(
+    publicKey: PublicKey,
+    callback: AccountChangeCallback,
+  ): number {
+    const id = ++this._accountChangeSubscriptionCounter;
+    this._accountChangeSubscriptions[id] = {
+      publicKey: publicKey.toBase58(),
+      callback,
+      subscriptionId: null,
+    };
+    this._updateSubscriptions();
+    return id;
+  }
+
+  /**
+   * Deregister an account notification callback
+   *
+   * @param id subscription id to deregister
+   */
+  async removeAccountChangeListener(id: number): Promise<void> {
+    if (this._accountChangeSubscriptions[id]) {
+      const {subscriptionId} = this._accountChangeSubscriptions[id];
+      delete this._accountChangeSubscriptions[id];
+      if (subscriptionId !== null) {
+        try {
+          await this._rpcWebSocket.call('accountUnsubscribe', [subscriptionId]);
+        } catch (err) {
+          console.log('accountUnsubscribe error:', err.message);
+        }
+      }
+      this._updateSubscriptions();
+    } else {
+      throw new Error(`Unknown account change id: ${id}`);
+    }
+  }
+
+  /**
+   * @private
+   */
+  _wsOnProgramAccountNotification(notification: Object) {
+    const res = ProgramAccountNotificationResult(notification);
+    if (res.error) {
+      throw new Error(res.error.message);
+    }
+
+    const keys = Object.keys(this._programAccountChangeSubscriptions).map(
+      Number,
+    );
+    for (let id of keys) {
+      const sub = this._programAccountChangeSubscriptions[id];
+      if (sub.subscriptionId === res.subscription) {
+        const {result} = res;
+        assert(typeof result !== 'undefined');
+
+        sub.callback({
+          accountId: result[0],
+          accountInfo: {
+            executable: result[1].executable,
+            owner: new PublicKey(result[1].owner),
+            // lamports: result[1].lamports,
+            dif: result[1].dif,
+            data: Buffer.from(result[1].data),
+          },
+        });
+        return true;
+      }
+    }
+  }
+
+  /**
+   * Register a callback to be invoked whenever accounts owned by the
+   * specified program change
+   *
+   * @param programId Public key of the program to monitor
+   * @param callback Function to invoke whenever the account is changed
+   * @return subscription id
+   */
+  onProgramAccountChange(
+    programId: PublicKey,
+    callback: ProgramAccountChangeCallback,
+  ): number {
+    const id = ++this._programAccountChangeSubscriptionCounter;
+    this._programAccountChangeSubscriptions[id] = {
+      programId: programId.toBase58(),
+      callback,
+      subscriptionId: null,
+    };
+    this._updateSubscriptions();
+    return id;
+  }
+
+  /**
+   * Deregister an account notification callback
+   *
+   * @param id subscription id to deregister
+   */
+  async removeProgramAccountChangeListener(id: number): Promise<void> {
+    if (this._programAccountChangeSubscriptions[id]) {
+      const {subscriptionId} = this._programAccountChangeSubscriptions[id];
+      delete this._programAccountChangeSubscriptions[id];
+      if (subscriptionId !== null) {
+        try {
+          await this._rpcWebSocket.call('programUnsubscribe', [subscriptionId]);
+        } catch (err) {
+          console.log('programUnsubscribe error:', err.message);
+        }
+      }
+      this._updateSubscriptions();
+    } else {
+      throw new Error(`Unknown account change id: ${id}`);
+    }
   }
 }
