@@ -1,7 +1,16 @@
 // @flow
 export {Account} from './account';
+export {BpfLoader} from './bpf-loader';
 export {BudgetProgram} from './budget-program';
 export {Connection} from './connection';
+export {Loader} from './loader';
+export {NativeLoader} from './native-loader';
 export {PublicKey} from './publickey';
 export {SystemProgram} from './system-program';
-export {Transaction} from './transaction';
+export {Token, TokenAmount} from './token-program';
+export {Transaction, TransactionInstruction} from './transaction';
+export {sendAndConfirmTransaction} from './util/send-and-confirm-transaction';
+export {
+  sendAndConfirmRawTransaction,
+} from './util/send-and-confirm-raw-transaction';
+export {testnetChannelEndpoint} from './util/testnet';
