@@ -1,9 +1,10 @@
 // @flow
 
-import assert from 'assert';
+import * as BufferLayout from 'buffer-layout';
 
 import {Transaction} from './transaction';
 import {PublicKey} from './publickey';
+import * as Layout from './layout';
 
 /**
  * Factory class for transactions to interact with the System program
@@ -13,7 +14,9 @@ export class SystemProgram {
    * Public key that identifies the System program
    */
   static get programId(): PublicKey {
-    return new PublicKey('0x000000000000000000000000000000000000000000000000000000000000000');
+    return new PublicKey(
+      '0x000000000000000000000000000000000000000000000000000000000000000',
+    );
   }
 
   /**
@@ -22,55 +25,63 @@ export class SystemProgram {
   static createAccount(
     from: PublicKey,
     newAccount: PublicKey,
-    tokens: number,
+    // lamports: number,
+    dif: number,
     space: number,
-    programId: PublicKey
+    programId: PublicKey,
   ): Transaction {
-    const userdata = Buffer.alloc(4 + 8 + 8 + 1 + 32);
-    let pos = 0;
+    const dataLayout = BufferLayout.struct([
+      BufferLayout.u32('instruction'),
+      // BufferLayout.ns64('lamports'),
+      BufferLayout.ns64('dif'),
+      BufferLayout.ns64('space'),
+      Layout.publicKey('programId'),
+    ]);
 
-    userdata.writeUInt32LE(0, pos); // Create Account instruction
-    pos += 4;
+    const data = Buffer.alloc(dataLayout.span);
+    dataLayout.encode(
+      {
+        instruction: 0, // Create Account instruction
+        // lamports,
+        dif,
+        space,
+        programId: programId.toBuffer(),
+      },
+      data,
+    );
 
-    userdata.writeUInt32LE(tokens, pos); // tokens as i64
-    pos += 8;
-
-    userdata.writeUInt32LE(space, pos); // space as u64
-    pos += 8;
-
-    const programIdBytes = programId.toBuffer();
-    programIdBytes.copy(userdata, pos);
-    pos += programIdBytes.length;
-
-    assert(pos <= userdata.length);
-
-    return new Transaction({
-      fee: 0,
-      keys: [from, newAccount],
+    return new Transaction().add({
+      keys: [
+        {pubkey: from, isSigner: true},
+        {pubkey: newAccount, isSigner: false},
+      ],
       programId: SystemProgram.programId,
-      userdata,
+      data,
     });
   }
 
   /**
-   * Generate a Transaction that moves tokens from one account to another
+   * Generate a Transaction that transfers lamports from one account to another
    */
-  static move(from: PublicKey, to: PublicKey, amount: number): Transaction {
-    const userdata = Buffer.alloc(4 + 8);
-    let pos = 0;
-    userdata.writeUInt32LE(2, pos); // Move instruction
-    pos += 4;
+  static transfer(from: PublicKey, to: PublicKey, amount: number): Transaction {
+    const dataLayout = BufferLayout.struct([
+      BufferLayout.u32('instruction'),
+      BufferLayout.ns64('amount'),
+    ]);
 
-    userdata.writeUInt32LE(amount, pos); // amount as u64
-    pos += 8;
+    const data = Buffer.alloc(dataLayout.span);
+    dataLayout.encode(
+      {
+        instruction: 2, // Move instruction
+        amount,
+      },
+      data,
+    );
 
-    assert(pos === userdata.length);
-
-    return new Transaction({
-      fee: 0,
-      keys: [from, to],
+    return new Transaction().add({
+      keys: [{pubkey: from, isSigner: true}, {pubkey: to, isSigner: false}],
       programId: SystemProgram.programId,
-      userdata,
+      data,
     });
   }
 
@@ -78,23 +89,24 @@ export class SystemProgram {
    * Generate a Transaction that assigns an account to a program
    */
   static assign(from: PublicKey, programId: PublicKey): Transaction {
-    const userdata = Buffer.alloc(4 + 32);
-    let pos = 0;
+    const dataLayout = BufferLayout.struct([
+      BufferLayout.u32('instruction'),
+      Layout.publicKey('programId'),
+    ]);
 
-    userdata.writeUInt32LE(1, pos); // Assign instruction
-    pos += 4;
+    const data = Buffer.alloc(dataLayout.span);
+    dataLayout.encode(
+      {
+        instruction: 1, // Assign instruction
+        programId: programId.toBuffer(),
+      },
+      data,
+    );
 
-    const programIdBytes = programId.toBuffer();
-    programIdBytes.copy(userdata, pos);
-    pos += programIdBytes.length;
-
-    assert(pos === userdata.length);
-
-    return new Transaction({
-      fee: 0,
-      keys: [from],
+    return new Transaction().add({
+      keys: [{pubkey: from, isSigner: true}],
       programId: SystemProgram.programId,
-      userdata,
+      data,
     });
   }
 }
