@@ -1,6 +1,4 @@
-module.exports = { // eslint-disable-line import/no-commonjs
-  'extends': [
-    "plugin:jest/recommended",
-    "../.eslintrc.js",
-  ]
+module.exports = {
+  // eslint-disable-line import/no-commonjs
+  extends: ['plugin:jest/recommended', '../.eslintrc.js'],
 };
