@@ -1,6 +1,6 @@
 // @flow
 
-import {Account, SystemProgram, BudgetProgram} from '../src';
+import {Account, BudgetProgram, SystemProgram} from '../src';
 
 test('createAccount', () => {
   const from = new Account();
@@ -20,36 +20,26 @@ test('createAccount', () => {
   // TODO: Validate transaction contents more
 });
 
-test('move', () => {
+test('transfer', () => {
   const from = new Account();
   const to = new Account();
   let transaction;
 
-  transaction = SystemProgram.move(
-    from.publicKey,
-    to.publicKey,
-    123,
-  );
+  transaction = SystemProgram.transfer(from.publicKey, to.publicKey, 123);
 
   expect(transaction.keys).toHaveLength(2);
   expect(transaction.programId).toEqual(SystemProgram.programId);
   // TODO: Validate transaction contents more
 });
 
-
 test('assign', () => {
   const from = new Account();
   const to = new Account();
   let transaction;
 
-  transaction = SystemProgram.assign(
-    from.publicKey,
-    to.publicKey,
-  );
+  transaction = SystemProgram.assign(from.publicKey, to.publicKey);
 
   expect(transaction.keys).toHaveLength(1);
   expect(transaction.programId).toEqual(SystemProgram.programId);
   // TODO: Validate transaction contents more
 });
-
-
